+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 0.3.0-dev
 ---------
 
+- Created client for frame extraction: :code:`cli_fext`.
 - Added option to disable export of frame0 topology in trajedit
 
 0.2.1 (2019-10-26)